requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
astor==0.8.1
attrs==24.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.18
botocore==1.35.18
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.4
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
duckdb==1.1.0
ecdsa==0.19.0
email_validator==2.2.0
Faker==19.13.0
fastapi==0.112.2
fastapi-cache2==0.2.2
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
holidays==0.56
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib_resources==6.4.4
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
mysql-connector-python==9.0.0
numpy==1.23.5
oauthlib==3.2.2
openai==1.43.1
packaging==24.1
pandas==1.5.3
pandasai==2.2.14
passlib==1.7.4
patsy==0.5.6
pendulum==3.0.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.3
pluggy==1.5.0
portalocker==2.10.1
prophet==1.1.5
proto-plus==1.24.0
protobuf==5.28.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydantic-settings==2.4.0
PyJWT==2.9.0
pymongo==4.8.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
redis==4.6.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.10.1
setuptools==65.5.0
simple-salesforce==1.12.6
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
SQLAlchemy==2.0.33
sqlglot==25.20.1
sqlglotrs==0.2.12
stanio==0.5.1
starlette==0.38.4
statsmodels==0.13.5
threadpoolctl==3.5.0
time-machine==2.15.0
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3<1.27,>=1.25.4
uvicorn==0.30.6
Werkzeug==3.0.4
zeep==4.2.1
pdfkit==1.0.0
reportlab==4.2.5